Zero all allocated pattern program memory
Pattern->prog is a pointer to an array of Instruction[].
Instruction is a union, which includes the struct Inst.
Setting the members of Inst will not necessarily initialise all of Instruction.
i.e. if sizeof(Inst) < sizeof(Instruction)
Any unallocated memory cause the memcmp() in Pattern_richcompare() to give
false negatives.

fixes #1